+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Update the `org.realityforge.arez` artifacts to version `0.204`.
 * Update the `org.realityforge.zemeckis` artifact to version `0.14`.
 * Update the `org.realityforge.grim` artifacts to version `0.09`.
+* Upgrade the `org.realityforge.proton` artifacts to version `0.57`.
+* Eliminate the dependency on `com.google.testing.compile:compile-testing` to avoid migration complexity when moving to a modern java version.
 
 ### [v0.188](https://github.com/react4j/react4j/tree/v0.188) (2022-02-10) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.187...v0.188)